refactor(blackjack): BlackjackResult에 결과를 모아서 저장하도록 변경
